requirements.txt
numpy
pandas
matplotlib.pyplot
seaborn